requirements.txt
absl-py==2.1.0
accelerate==0.28.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
aliyun-python-sdk-core==2.15.0
aliyun-python-sdk-kms==2.16.2
altair==5.3.0
anndata==0.10.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
APScheduler==3.10.4
array_api_compat==1.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
av==11.0.0
beartype==0.18.2
blessed==1.20.0
build==1.2.1
CacheControl==0.14.0
cached-property==1.5.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
clean-fid==0.1.35
cleo==2.1.0
click==8.1.7
cmake==3.25.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.2
contourpy==1.2.0
crashtest==0.4.1
crcmod==1.7
cryptography==42.0.5
cycler==0.12.1
Cython==3.0.10
datasets==2.18.0
debugpy==1.8.2
decorator==5.1.1
diart==0.9.0
dill==0.3.8
distlib==0.3.8
dnspython==2.6.1
docker-pycreds==0.4.0
docopt==0.6.2
dulwich==0.21.7
einops==0.7.0
email_validator==2.1.1
exceptiongroup==1.2.1
executing==2.0.1
fast-bss-eval==0.1.4
fastapi==0.111.0
fastapi-cli==0.0.4
faster-whisper==1.0.1
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
gpustat==1.1.1
gradio==4.31.5
gradio_client==0.16.4
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
installer==0.7.0
InstructorEmbedding==1.0.1
ipykernel==6.29.5
ipywidgets==8.1.3
jaraco.classes==3.4.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.3
jiwer==3.0.3
jmespath==0.10.0
joblib==1.3.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
julius==0.2.7
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
keyring==24.3.1
kiwisolver==1.4.5
lazy_loader==0.3
legacy-api-wrap==1.4
librosa==0.10.2.post1
lightning==2.1.4
lightning-bolts==0.7.0
lightning-utilities==0.10.1
lit==15.0.7
llvmlite==0.42.0
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mir-eval==0.7
modelscope==1.13.3
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
multiprocess==0.70.16
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
omegaconf==2.3.0
openai-whisper==20231117
opencv-python==4.9.0.80
optuna==3.5.0
orjson==3.10.3
oss2==2.18.4
packaging==23.2
pandas==2.0.3
parso==0.8.4
pastel==0.2.1
patsy==0.5.6
pb-bss-eval==0.0.2
pesq==0.0.4
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.11.0
platformdirs==4.2.0
poethepoet==0.10.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
pooch==1.8.0
POT==0.9.4
primePy==1.3
prompt_toolkit==3.0.47
protobuf==4.25.2
psutil==5.9.8
ptflops==0.7.2.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycocotools==2.0.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.7.2
pydantic_core==2.18.3
pydub==0.25.1
Pygments==2.17.2
pynndescent==0.5.13
pyparsing==3.1.1
pyproject_hooks==1.1.0
pystoi==0.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==1.9.5
pytorch-metric-learning==2.4.1
pytorch-ranger==0.1.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.8.1
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
resampy==0.4.3
rich==13.7.0
rotary-embedding-torch==0.5.3
rpds-py==0.18.1
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
ruff==0.4.6
Rx==3.2.0
safetensors==0.4.2
scanpy==1.10.2
scikit-learn==1.4.0
scipy==1.12.0
scprep==1.2.3
seaborn==0.13.2
SecretStorage==3.3.3
semantic-version==2.10.0
semver==3.0.2
sentence-transformers==2.6.1
sentencepiece==0.1.99
sentry-sdk==1.41.0
session-info==1.0.0
setproctitle==1.3.3
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.3.7
speechbrain==0.5.16
SQLAlchemy==2.0.25
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.2
stdlib-list==0.10.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tiktoken==0.6.0
tokenizers==0.15.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch-audiomentations==0.11.0
torch-complex==0.4.3
torch-mir-eval==0.4
torch-optimizer==0.1.0
torch-pitch-shift==1.2.4
torch-stoi==0.2.1
torchcde==0.2.5
torchcfm==1.0.5
torchdiffeq==0.2.4
torchdyn==1.0.6
torchmetrics==0.11.4
torchsde==0.2.6
tornado==6.4.1
tqdm==4.66.1
traitlets==5.14.3
trampoline==0.1.2
transformers==4.39.2
triton==2.0.0
trove-classifiers==2024.5.22
typeguard==2.13.3
typer==0.12.3
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ujson==5.10.0
umap-learn==0.5.6
urllib3==2.2.0
uvicorn==0.30.0
uvloop==0.19.0
virtualenv==20.26.2
wandb==0.16.4
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.7.0
websocket-server==0.6.4
websockets==11.0.3
Werkzeug==3.0.3
widgetsnbextension==4.0.11
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1